# Speed up `OnDigraphs` for a digraph and a permutation