--- a/gap/oper.gd
+++ b/gap/oper.gd
@@ -62,6 +62,8 @@ DeclareOperation("DIGRAPHS_GraphProduct", [IsDigraph, IsDigraph, IsFunction]);
 # 4. Actions . . .
 DeclareOperation("OnDigraphs", [IsDigraph, IsPerm]);
 DeclareOperation("OnDigraphs", [IsDigraph, IsTransformation]);
+DeclareOperation("OnDigraphsNC", [IsDigraph, IsPerm]);
+DeclareOperation("OnDigraphsNC", [IsDigraph, IsTransformation]);
 DeclareOperation("OnTuplesDigraphs", [IsDigraphCollection, IsPerm]);
 DeclareOperation("OnSetsDigraphs", [IsDigraphCollection, IsPerm]);
 DeclareOperation("OnMultiDigraphs", [IsDigraph, IsPermCollection]);