--- a/gap/oper.gi
+++ b/gap/oper.gi
@@ -993,17 +993,23 @@ end);
 # 4. Actions
 ###############################################################################
 
-InstallMethod(OnDigraphs, "for a mutable digraph by out-neighbours and a perm",
-[IsMutableDigraph and IsDigraphByOutNeighboursRep, IsPerm],
+InstallMethod(OnDigraphs, "for a digraph and a perm",
+[IsDigraph, IsPerm],
 function(D, p)
-  local out;
-  if SmallestMovedPoint(p) > DigraphNrVertices(D) then
+  if ForAll(DigraphVertices(D), i -> i ^ p = i) then
     return D;
   elif ForAny(DigraphVertices(D), i -> i ^ p > DigraphNrVertices(D)) then
     ErrorNoReturn("the 2nd argument <p> must be a permutation that permutes ",
                   "the vertices of the digraph <D> that is the 1st argument");
   fi;
+  return OnDigraphsNC(D, p);
+end);
 
+InstallMethod(OnDigraphsNC,
+"for a mutable digraph by out-neighbours and a perm",
+[IsMutableDigraph and IsDigraphByOutNeighboursRep, IsPerm],
+function(D, p)
+  local out;
   out := D!.OutNeighbours;
   out{DigraphVertices(D)} := Permuted(out, p);
   Apply(out, x -> OnTuples(x, p));
@@ -1011,26 +1017,38 @@ function(D, p)
   return D;
 end);
 
-InstallMethod(OnDigraphs, "for a immutable digraph and a perm",
+InstallMethod(OnDigraphsNC, "for a immutable digraph and a perm",
 [IsImmutableDigraph, IsPerm],
 function(D, p)
-  if SmallestMovedPoint(p) > DigraphNrVertices(D) then
-    return D;
-  fi;
-  return MakeImmutable(OnDigraphs(DigraphMutableCopy(D), p));
+  local out, permed;
+  out := D!.OutNeighbours;
+  permed := Permuted(out, p);
+  Apply(permed, x -> OnTuples(x, p));
+  return DigraphNC(IsImmutableDigraph, permed);
 end);
 
 InstallMethod(OnDigraphs,
-"for a mutable digraph by out-neighbours and a transformation",
-[IsMutableDigraph and IsDigraphByOutNeighboursRep, IsTransformation],
+"for a digraph and a transformation",
+[IsDigraph, IsTransformation],
 function(D, t)
   local old, new, v;
-  if SmallestMovedPoint(t) > DigraphNrVertices(D) then
+  if ForAll(DigraphVertices(D), i -> i ^ t = i) then
     return D;
   elif ForAny(DigraphVertices(D), i -> i ^ t > DigraphNrVertices(D)) then
     ErrorNoReturn("the 2nd argument <t> must be a transformation that ",
                   "maps every vertex of the digraph <D> that is the 1st ",
                   "argument, to another vertex");
+  fi;
+  return OnDigraphsNC(D, t);
+end);
+
+InstallMethod(OnDigraphsNC,
+"for a mutable digraph by out-neighbours and a transformation",
+[IsMutableDigraph and IsDigraphByOutNeighboursRep, IsTransformation],
+function(D, t)
+  local old, new, v;
+  if t = IdentityTransformation then
+    return D;
   fi;
   old := D!.OutNeighbours;
   new := List(DigraphVertices(D), x -> []);
@@ -1042,13 +1060,19 @@ function(D, t)
   return D;
 end);
 
-InstallMethod(OnDigraphs, "for a immutable digraph and a transformation",
+InstallMethod(OnDigraphsNC, "for a immutable digraph and a transformation",
 [IsImmutableDigraph, IsTransformation],
 function(D, t)
-  if SmallestMovedPoint(t) > DigraphNrVertices(D) then
+  local old, new, v;
+  if t = IdentityTransformation then
     return D;
   fi;
-  return MakeImmutable(OnDigraphs(DigraphMutableCopy(D), t));
+  old := D!.OutNeighbours;
+  new := List(DigraphVertices(D), x -> []);
+  for v in DigraphVertices(D) do
+    Append(new[v ^ t], OnTuples(old[v], t));
+  od;
+  return DigraphNC(IsImmutableDigraph, new);
 end);
 
 InstallMethod(OnTuplesDigraphs,