@wilfwilson 2019-10-28
Another solution is that we could create some `NC` versions of functions that are particularly performance-critical, including `OnDigraphsNC`.