@james-d-mitchell 2025-04-07
I'd propose improving the performance of `OnDigraphs` at gap days in Brussels in 2025, if we have the time.