@reiniscirpons 2025-04-10
I tried my best to incorporate the improved methods, but I just do not see the timing improvements claimed. I settled on implementing a separate `OnDigraphsNC` method which skips the more expensive `ForAll` and `ForEach` checks and only does the `p = ()` optimization. I then replaced internal usage of the `OnDigraphs` method with `OnDigraphsNC`. Here are some runtime stats:

|Test |Old method time, ms|Proposed new `OnDigraphsNC` method time, ms| Proposed new `OnDigraphs` method time, ms
|-|-|-|-|
| 1000 random digraphs with 100 vertices, `IsIsomorphicDigraph` | 703 | **689** | - |
| 829 random topologically sortable digraphs with 100 vertices, `DigraphTransitiveReduction` | 65 | **52** | - |
|100 random transitive digraphs with 1000 vertices, `IsIsomorphicDigraph` | **5476** | 5374 | - |
|100 random transitive digraphs with 1000 vertices, `DigraphTransitiveReduction` | 24020 | **23947** | - |
|100 random digraphs with 1000 vertices, `OnDigraphs` or `OnDigraphsNC` respectively | 1415 | **1403** | 1487 |

There seem to be some minor improvement (within margin of error in some cases) with the proposed `OnDigraphsNC` method, but it slows down `OnDigraphs` by a bit.